# docs/api/swagger.yaml
openapi: 3.0.0
info:
  title: Restaurant Management API
  version: 1.0.0
  description: |
    ### API для управления рестораном

    Этот API предоставляет доступ к данным ресторана, включая управление сотрудниками, клиентами, заказами и поставками. 
    Пользователи могут выполнять такие действия, как поиск блюд по ингредиентам, расчет средней стоимости заказов и получение информации о поставках.

paths:
  /api/dishes/get:
    get:
      tags:
        - Dishes
      summary: Список блюд и их ингредиентов
      operationId: getDishesWithIngredients
      responses:
        '200':
          description: Список блюд и их ингредиентов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DishWithIngredients'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/dishes/get/unordered:
    get:
      tags:
        - Dishes
      summary: Список блюд и их ингредиентов
      operationId: getUnorderedDishes
      responses:
        '200':
          description: Список блюд и их ингредиентов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dish'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/dishes/get/by-ingredients:
    get:
      tags:
        - Dishes
      summary: Поиск блюд по ингредиентам
      operationId: getDishesByIngredients
      description: |
        Используйте `ingredients` как query параметр для поиска.
        Пример: `/api/dishes/get/by-ingredients?ingredient=tomato`
      responses:
        '200':
          description: Список блюд, содержащих указанные ингредиенты
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dish'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/employees/orders-count:
    get:
      tags:
        - Employees
      summary: Количество заказов, выполненных каждым сотрудником за период
      operationId: getEmployeeOrderCount
      description: |
        Используйте `date_from` и `date_to` как query параметры для фильтрации по периоду.
        Пример: `/api/employees/orders-count?date_from=2023-01-01&date_to=2023-12-31`
      responses:
        '200':
          description: Количество заказов, выполненных каждым сотрудником
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeOrderCount'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/products/expiring-soon:
    get:
      tags:
        - Supplies
      summary: Получение продуктов с истекающим сроком годности
      operationId: getExpiringProducts
      description: Получение продуктов с истекающим сроком годности
      responses:
        '200':
          description: Список продуктов с истекающим сроком годности
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/report/download:
    get:
      tags:
        - Report download
      summary: Скачать отчет по финансам
      operationId: downloadReport
      description: |
        Используйте `date_from` и `date_to` как query параметры для фильтрации по периоду.
        Пример: `/employees/orders-count?date_from=2023-01-01&date_to=2023-12-31`
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: byte
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/customers/{customer_id}/orders:
    get:
      tags:
        - Customers
      summary: Получение истории заказов клиента
      operationId: getCustomerOrderHistory
      description: |
        Используйте query параметры `date_from` и `date_to` для фильтрации истории заказов по периоду.
        Пример: `/api/customers/{customer_id}/orders?date_from=2023-01-01&date_to=2023-12-31`
      parameters:
        - name: customer_id
          in: path
          required: true
          schema:
            type: integer
          description: Идентификатор клиента.
      responses:
        '200':
          description: История заказов клиента
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerOrderHistory'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Клиент не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/dishes/popular:
    get:
      tags:
        - Dishes
      summary: Список самых популярных блюд
      operationId: getMostPopularDishes
      responses:
        '200':
          description: Список самых популярных блюд
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DishWithIngredients'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/products/{product_name}/suppliers:
    get:
      summary: Получение всех поставщиков для определенного продукта
      operationId: getSuppliersByProduct
      parameters:
        - name: product_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Список поставщиков, которые поставляют указанный продукт
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplierWithSupplies'
        '400':
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Продукт не найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Продукт не найден

components:
  schemas:
    Employee:
      type: object
      properties:
        employee_id:
          type: integer
          description: Идентификатор сотрудника
        first_name:
          type: string
          description: Имя сотрудника
        last_name:
          type: string
          description: Фамилия сотрудника
        position_name:
          type: string
          description: Название позиции
        salary:
          type: number
          format: double
          description: Зарплата сотрудника

    MostPopularDish:
      type: object
      properties:
        dish_id:
          type: integer
          description: Идентификатор блюда
        name:
          type: string
          description: Название блюда
        price:
          type: number
          format: double
          description: Цена блюда
        ordered_count:
          type: integer
          format: int64
          description: количество раз заказано

    Dish:
      type: object
      properties:
        dish_id:
          type: integer
          description: Идентификатор блюда
        name:
          type: string
          description: Название блюда
        price:
          type: number
          format: double
          description: Цена блюда

    Supplier:
      type: object
      properties:
        supplier_id:
          type: integer
          description: Идентификатор поставщика
        company_name:
          type: string
          description: Название компании-поставщика
        chief_name:
          type: string
        stock_quantity:
          type: integer
          description: Количество товара на складе

    Product:
      type: object
      properties:
        product_id:
          type: integer
        name:
          type: string
        date_of_production:
          type: string
          format: date
        date_of_expiry:
          type: string
          format: date
        status:
          type: string

    EmployeeOrderCount:
      type: object
      properties:
        employee_id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        orders_count:
          type: integer

    DishWithIngredients:
      type: object
      properties:
        dish_id:
          type: integer
        name:
          type: string
        price:
          type: number
          format: double
        ingredients:
          type: array
          items:
            type: string

    CustomerOrderHistory:
      type: object
      properties:
        customer_id:
          type: integer
          description: Идентификатор клиента
        first_name:
          type: string
        last_name:
          type: string
        orders:
          type: array
          items:
            type: object
            properties:
              order_id:
                type: integer
                description: Идентификатор заказа
              created_at:
                type: string
                format: date
                description: Дата и время создания заказа
              total_price:
                type: number
                format: double
                description: Общая стоимость заказа
              dishes:
                type: array
                items:
                  type: object
                  properties:
                    dish_id:
                      type: integer
                      description: Идентификатор блюда
                    name:
                      type: string
                      description: Название блюда
                    quantity:
                      type: integer
                      description: Количество блюда в заказе
                    price:
                      type: number
                      format: double
                      description: Цена блюда на момент заказа

    SupplierWithSupplies:
      type: object
      properties:
        supplier_id:
          type: integer
        company_name:
          type: string
        supplies:
          type: array
          items:
            type: object
            properties:
              supply_id:
                type: integer
              date_from:
                type: string
                format: date-time
              price:
                type: number
                format: double
              quantity:
                type: integer

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Код ошибки
        message:
          type: string
          description: Сообщение об ошибке